Change: remove resolv.conf if it didn't exist previously
Note: If the resolv.conf didn't exist previously, we consider the freshly
created resolv.conf as private to gof5 and use mode 0644, and remove it
on exit.

(The 0644 is still subject to umask. If you've the strong feeling, that
0666+umask is the right value, please contact me.)

If it exists, we do not touch the owner and permissions, but the
content.